Added the first iteration of synthetic classes.
These allow you to implement interfaces and/or extend Java classes from within Python, then instantiate them, and pass them to Java methods which executes the Python code. The first iteration is a proof of concept and does work although has serious limitations including only supporting classes that make no use of self, the classes have strict requirements on them, and the function call to create the synthetic class is cumbersome. There are plans to remedy all of these problems now that the 'interesting' parts are solved.